Update 2.3 fork

+++ dev/tests/integration/testsuite/Magento/ConfigurableProduct/_files/product_configurable_rollback.php
@@ -18,13 +18,15 @@ $productRepository = \Magento\TestFramework\Helper\Bootstrap::getObjectManager()
 
 foreach (['simple_10', 'simple_20', 'configurable'] as $sku) {
     try {
-        $product = $productRepository->get($sku, false, null, true);
+        $product = $productRepository->get($sku, true);
 
         $stockStatus = $objectManager->create(\Magento\CatalogInventory\Model\Stock\Status::class);
         $stockStatus->load($product->getEntityId(), 'product_id');
         $stockStatus->delete();
 
-        $productRepository->delete($product);
+        if ($product->getId()) {
+            $productRepository->delete($product);
+        }
     } catch (\Magento\Framework\Exception\NoSuchEntityException $e) {
         //Product already removed
     }
